cleaned up code. not finished yet
