+++ DodgerGame/src/main/java/Controller/KeyController.java
@@ -1,27 +1,16 @@
 package Controller;
 
-import Interfaces.IGameWorldObserver;
 import Model.GameWorld;
-import Model.PausableAnimationTimer;
-import View.PauseMenu;
-import Model.Entities.Player.Spaceship;
 import javafx.scene.input.KeyEvent;
 import javafx.stage.Stage;
 
-public class KeyController implements IGameWorldObserver {
+public class KeyController {
 
+    // Stage is pausing the simulation
     private final Stage stage;
-    private GameWorld gameWorld;
-    private Spaceship spaceship;
-    private final PausableAnimationTimer gameLoop;
-    private final PauseMenu pauseMenu;
 
-    public KeyController(Stage stage, PausableAnimationTimer gameLoop, PauseMenu pauseMenu) {
-        this.gameLoop = gameLoop;
+    public KeyController(Stage stage) {
         this.stage = stage;
-        this.gameWorld = GameWorld.getInstance();
-        this.spaceship = gameWorld.getSpaceship();
-        this.pauseMenu = pauseMenu;
     }
 
     /**
@@ -33,16 +22,16 @@ public class KeyController implements IGameWorldObserver {
     public void handleKeyPressed(KeyEvent event) {
         switch (event.getCode()) {
             case UP:
-                spaceship.setUp(1);
+                GameWorld.getInstance().getSpaceship().setUp(1);
                 break;
             case DOWN:
-                spaceship.setDown(1);
+                GameWorld.getInstance().getSpaceship().setDown(1);
                 break;
             case LEFT:
-                spaceship.setLeft(1);
+                GameWorld.getInstance().getSpaceship().setLeft(1);
                 break;
             case RIGHT:
-                spaceship.setRight(1);
+                GameWorld.getInstance().getSpaceship().setRight(1);
                 break;
             case ESCAPE:
                 // TODO Implement pause function and pause menu
@@ -63,25 +52,19 @@ public class KeyController implements IGameWorldObserver {
     public void handleKeyReleased(KeyEvent event) {
         switch (event.getCode()) {
             case UP:
-                spaceship.setUp(0);
+                GameWorld.getInstance().getSpaceship().setUp(0);
                 break;
             case DOWN:
-                spaceship.setDown(0);
+                GameWorld.getInstance().getSpaceship().setDown(0);
                 break;
             case LEFT:
-                spaceship.setLeft(0);
+                GameWorld.getInstance().getSpaceship().setLeft(0);
                 break;
             case RIGHT:
-                spaceship.setRight(0);
+                GameWorld.getInstance().getSpaceship().setRight(0);
                 break;
             default:
                 break;
         }
-    }
-
-    @Override
-    public void actOnEvent() {
-        gameWorld = GameWorld.getInstance();
-        spaceship = gameWorld.getSpaceship();
     }
 }
