--- DodgerGame/src/main/java/Interfaces/IGameWorldObservable.java
@@ -1,10 +1,0 @@
-package Interfaces;
-
-public interface IGameWorldObservable {
-
-    public void notifyGameWorldObservers();
-
-    public void addGameWorldObserver(IGameWorldObserver obs);
-
-    public void removeGameWorldObserver(IGameWorldObserver obs);
-}
