--- DodgerGame/src/main/java/Interfaces/IGameWorldObserver.java
@@ -1,5 +1,0 @@
-package Interfaces;
-
-public interface IGameWorldObserver {
-    public void actOnEvent();
-}
