+++ DodgerGame/src/main/java/View/HealthBarGUI.java
@@ -11,7 +11,7 @@ import java.io.InputStream;
  * @author Viktor Sundberg (viktor.sundberg@icloud.com)
  */
 
-public class HealthBarGUI implements ISpaceshipObserver {
+public class HealthBarGUI {
     private final GraphicsContext gc;
     private Image background;
     private Image foreground;
@@ -55,14 +55,9 @@ public class HealthBarGUI implements ISpaceshipObserver {
      * @param maxHealth the maximum amount of health point
      * @author Olle Westerlund
      */
-    private void drawHealthBar(double remainingHp, double maxHealth) {
+    public void drawHealthBar(double remainingHp, double maxHealth) {
         gc.drawImage(background, 0, 0, maxHealth, 40);
         gc.drawImage(foreground, 0, 0, remainingHp, 40);
         gc.drawImage(border, 0, 0, maxHealth, 40);
-    }
-
-    @Override
-    public void actOnEvent(Spaceship spaceship) {
-        drawHealthBar(spaceship.getHp(), spaceship.getMaxHp());
     }
 }
