+++ DodgerGame/src/main/java/View/Sound/SoundHandler.java
@@ -12,7 +12,7 @@ import java.io.File;
  * @Author Viktor Sundberg (viktor.sundberg@icloud.com)
  */
 
-public class SoundHandler implements ISoundObserve {
+public class SoundHandler {
 
     private double volume = 0.01;
 
@@ -64,8 +64,7 @@ public class SoundHandler implements ISoundObserve {
         gain.setValue(decibel);
     }
 
-    @Override
-    public void actOnEvent(Class c) {
+    public void playSound(Class c) {
         String soundFilepath = null;
 
         if (c.equals(Asteroid.class)) {
