+++ DodgerGame/src/main/resources/HighScores.txt
@@ -1,0 +1,2 @@
+42
+39
